AutoAffiliateHub-X2/requirements.txt
# StockSpot - Amazon + Twitter/X Only Dependencies
# 
# Quick Installation:
# pip install -r requirements.txt

# Core Web Framework
Flask>=3.0.0

# Twitter/X Integration
tweepy>=4.14.0

# HTTP Requests & Environment
requests>=2.32.0
python-dotenv>=1.0.0

# Amazon Product Advertising API
python-amazon-paapi>=5.0.0

# HTML Parsing for Link Processing  
beautifulsoup4>=4.12.0

# Development & Testing
pytest>=7.4.0
pytest-flask>=1.3.0

# Data processing
python-dateutil>=2.8.0
pytz>=2023.3

# Development and testing
pytest>=7.4.0
pytest-flask>=1.3.0

# Production server
gunicorn>=21.2.0

# Optional: Enhanced security
Flask-WTF==1.2.1
WTForms==3.1.1

# Optional: Caching
Flask-Caching==2.1.0